[INJIMOB-1433] refactor
